Create form to select age
