+++ components/Styled-Components/AdultNext.jsx
@@ -6,11 +6,13 @@ import styled from "styled-components";
 // ask maria if you are confused (:
 
 const AdultNext = styled.button`
-position: absolute;
-top: 50%;
-left: 50%;
-margin-top: -7rem;
-margin-left: -7.5rem;
+
+display: flex;
+text-align: center;
+justify-content: center;
+margin-left: auto;
+margin-right: auto;
+margin-top: 10rem;
 width: 13rem;
 padding: 1.5rem;
 box-shadow: 5px 10px 8px #888888;
@@ -20,7 +22,6 @@ border-radius: 20px;
 border: none;
 background-color: #4316A3;
 letter-spacing: 1px;
-text-align: center;
 cursor: pointer;
 
 &:hover{
