+++ pages/adult-onboarding2.js
@@ -1,28 +1,20 @@
-import Link from "next/link";
-import styled from "styled-components";
 import AdultsBack from "../components/AdultsBack";
 import AdultsNext from "../components/AdultsNext";
 import AdultText from "../components/Styled-Components/AdultText";
+import Title from "../components/Styled-Components/Title";
 
 export default function Adult() {
-  const getAge = (e) => {
-    e.preventDefault();
-    // picking up the age.
-    // console.log(e.target.age.value);
-    let learnersAge = e.target.age.value;
-    console.log(learnersAge);
-    // here we will handle saving the users age to local storage.
-  };
   return (
+    // crab claws as a parent container
+    // sending it the page it is coming from for back / next functionality
+    // if its the first page we will send different text to the button
+    // will accept a boolean of true.
+
+    // &apos; is html code for apostrophe (not to break depoly)
     <>
-      <AdultText>How old is the Learner?</AdultText>
-      <form onSubmit={getAge}>
-        <label>
-          <input type="number" name="age" min="1" max="18"></input>
-          years
-        </label>
-        <AdultsNext page={2} />
-      </form>
+      <Title>Welcome!</Title>
+      <AdultText>Let&apos;s set up a profile for the learner.</AdultText>
+      <AdultsNext page={2} />
       <AdultsBack page={2} />
     </>
   );
